(add) interfaces file to modularize commonly used interfaces
